+++ packages/encryption/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.19.24](https://github.com/standardnotes/app/compare/@standardnotes/encryption@1.19.23...@standardnotes/encryption@1.19.24) (2022-12-02)
+
+**Note:** Version bump only for package @standardnotes/encryption
+
 ## [1.19.23](https://github.com/standardnotes/app/compare/@standardnotes/encryption@1.19.22...@standardnotes/encryption@1.19.23) (2022-12-01)
 
 **Note:** Version bump only for package @standardnotes/encryption